+++ packages/integration-tests/__tests__/defined-behavior/2-expected-use/ai-conversation.ts
@@ -7,6 +7,7 @@ import {
   subOptionsAndHeaders,
   pause,
 } from '../../utils';
+import { Subscriber } from 'rxjs';
 
 describe('AI Conversation Routes', () => {
   // data/resource.ts
@@ -64,7 +65,7 @@ describe('AI Conversation Routes', () => {
         listMessages: expect.any(Function),
         metadata: {},
         name: 'Test Conversation',
-        onMessage: expect.any(Function),
+        onStreamEvent: expect.any(Function),
         sendMessage: expect.any(Function),
         updatedAt: '2023-08-02T12:00:00Z',
       });
@@ -113,7 +114,56 @@ describe('AI Conversation Routes', () => {
         listMessages: expect.any(Function),
         metadata: {},
         name: 'Test Conversation',
-        onMessage: expect.any(Function),
+        onStreamEvent: expect.any(Function),
+        sendMessage: expect.any(Function),
+        updatedAt: '2023-08-02T12:00:00Z',
+      });
+      // #endregion assertions
+    });
+
+    test('Update a conversation', async () => {
+      const sampleConversation = {
+        id: 'conversation-id',
+        createdAt: '2023-06-01T12:00:00Z',
+        updatedAt: '2023-08-02T12:00:00Z',
+        metadata: {},
+        name: 'Test Conversation',
+      };
+
+      const { spy, generateClient } = mockedGenerateClient([
+        {
+          data: {
+            updateConversation: sampleConversation,
+          },
+        },
+      ]);
+      // simulated amplifyconfiguration.json
+      const config = await buildAmplifyConfig(schema);
+      // #endregion mocking
+
+      // #region api call
+      // App.tsx
+      Amplify.configure(config);
+      const client = generateClient<Schema>();
+      // create conversation
+      const { data: updatedConversation, errors: updateConversationErrors } =
+        await client.conversations.chatBot.update({
+          id: sampleConversation.id,
+          name: 'updated conversation name',
+          metadata: { arbitrary: 'data' },
+        });
+      // #endregion api call
+
+      // #region assertions
+      expect(optionsAndHeaders(spy)).toMatchSnapshot();
+      expect(updateConversationErrors).toBeUndefined();
+      expect(updatedConversation).toStrictEqual({
+        createdAt: '2023-06-01T12:00:00Z',
+        id: sampleConversation.id,
+        listMessages: expect.any(Function),
+        metadata: {},
+        name: 'Test Conversation',
+        onStreamEvent: expect.any(Function),
         sendMessage: expect.any(Function),
         updatedAt: '2023-08-02T12:00:00Z',
       });
@@ -156,7 +206,7 @@ describe('AI Conversation Routes', () => {
         listMessages: expect.any(Function),
         metadata: {},
         name: 'Test Conversation',
-        onMessage: expect.any(Function),
+        onStreamEvent: expect.any(Function),
         sendMessage: expect.any(Function),
         updatedAt: '2023-06-01T12:00:00Z',
       });
@@ -209,7 +259,7 @@ describe('AI Conversation Routes', () => {
           listMessages: expect.any(Function),
           metadata: {},
           name: 'Test Conversation',
-          onMessage: expect.any(Function),
+          onStreamEvent: expect.any(Function),
           sendMessage: expect.any(Function),
           updatedAt: '2023-08-02T12:00:00Z',
         },
@@ -219,7 +269,7 @@ describe('AI Conversation Routes', () => {
           listMessages: expect.any(Function),
           metadata: {},
           name: 'Test Conversation2',
-          onMessage: expect.any(Function),
+          onStreamEvent: expect.any(Function),
           sendMessage: expect.any(Function),
           updatedAt: '2024-09-05T12:00:00Z',
         },
@@ -280,7 +330,7 @@ describe('AI Conversation Routes', () => {
           listMessages: expect.any(Function),
           metadata: {},
           name: 'Test Conversation',
-          onMessage: expect.any(Function),
+          onStreamEvent: expect.any(Function),
           sendMessage: expect.any(Function),
           updatedAt: '2023-08-02T12:00:00Z',
         },
@@ -290,7 +340,7 @@ describe('AI Conversation Routes', () => {
           listMessages: expect.any(Function),
           metadata: {},
           name: 'Test Conversation2',
-          onMessage: expect.any(Function),
+          onStreamEvent: expect.any(Function),
           sendMessage: expect.any(Function),
           updatedAt: '2024-09-05T12:00:00Z',
         },
@@ -318,6 +368,45 @@ describe('AI Conversation Routes', () => {
       createdAt: '2024-08-22T18:28:00.596Z',
       id: 'message-id',
       role: 'user',
+    };
+    const sampleConversationStreamTextEvent = {
+      id: 'stream-text-event-id',
+      conversationId: sampleConversation.id,
+      associatedUserMessageId: sampleConversationMessage1.id,
+      contentBlockDeltaIndex: 0,
+      contentBlockIndex: 0,
+      contentBlockText: 'foo',
+    };
+    const sampleConversationStreamToolUseEvent = {
+      id: 'stream-tooluse-event-id',
+      conversationId: sampleConversation.id,
+      associatedUserMessageId: sampleConversationMessage1.id,
+      contentBlockIndex: 0,
+      contentBlockToolUse: {
+        toolUseId: 'toolUseId',
+        name: 'toolUseName',
+        input: JSON.stringify({ toolUseParam: 'toolUseParam' }),
+      },
+    };
+    const sampleConversationStreamDoneAtIndexEvent = {
+      id: 'stream-doneatindex-event-id',
+      conversationId: sampleConversation.id,
+      associatedUserMessageId: sampleConversationMessage1.id,
+      contentBlockDoneAtIndex: 0,
+      contentBlockIndex: 0,
+    };
+    const sampleConversationStreamTurnDoneEvent = {
+      id: 'stream-turndone-event-id',
+      conversationId: sampleConversation.id,
+      associatedUserMessageId: sampleConversationMessage1.id,
+      contentBlockIndex: 0,
+      stopReason: 'stopReason',
+    };
+    const sampleConversationStreamErrorEvent = {
+      id: 'stream-error-event-id',
+      conversationId: sampleConversation.id,
+      associatedUserMessageId: sampleConversationMessage1.id,
+      errors: [{ message: 'error message', errorType: 'errorType' }],
     };
     // #endregion mocking common
 
@@ -471,6 +560,117 @@ describe('AI Conversation Routes', () => {
       // #endregion assertions
     });
 
+    describe('Stream events', () => {
+      const mockNextHandler = jest.fn();
+      const mockErrorHandler = jest.fn();
+      let subs: Record<string, Subscriber<any>>;
+
+      beforeAll(async () => {
+        const { subs: mockedSubs, generateClient } = mockedGenerateClient([
+          {
+            data: {
+              getConversation: sampleConversation,
+            },
+          },
+        ]);
+        subs = mockedSubs;
+
+        const config = await buildAmplifyConfig(schema);
+        Amplify.configure(config);
+
+        const client = generateClient<Schema>();
+        const { data: conversation } = await client.conversations.chatBot.get({
+          id: sampleConversation.id,
+        });
+        // subscribe to messages
+        conversation?.onStreamEvent({
+          next: (streamEvent) => mockNextHandler(streamEvent),
+          error: (error) => mockErrorHandler(error),
+        });
+      });
+
+      test('Text event', async () => {
+        subs.onCreateAssistantResponseChatBot.next({
+          data: {
+            onCreateAssistantResponseChatBot: {
+              ...sampleConversationStreamTextEvent,
+            },
+          },
+        });
+
+        await pause(1);
+        const {
+          contentBlockText,
+          ...rest
+        } = sampleConversationStreamTextEvent;
+        const expectedConversationStreamEvent = {
+          text: contentBlockText,
+          ...rest,
+        };
+        expect(mockNextHandler).toHaveBeenCalledWith(expectedConversationStreamEvent);
+      });
+
+      test('Tool use event', async () => {
+        subs.onCreateAssistantResponseChatBot.next({
+          data: {
+            onCreateAssistantResponseChatBot: {
+              ...sampleConversationStreamToolUseEvent,
+            },
+          },
+        });
+
+        await pause(1);
+        const {
+          contentBlockToolUse: { input, toolUseId, name },
+          ...rest
+        } = sampleConversationStreamToolUseEvent;
+        const expectedConversationStreamEvent = {
+          toolUse: { input: JSON.parse(input), toolUseId, name },
+          ...rest,
+        };
+        expect(mockNextHandler).toHaveBeenCalledWith(expectedConversationStreamEvent);
+      });
+
+      test('Done at index event', async () => {
+        subs.onCreateAssistantResponseChatBot.next({
+          data: {
+            onCreateAssistantResponseChatBot: {
+              ...sampleConversationStreamDoneAtIndexEvent,
+            },
+          },
+        });
+
+        await pause(1);
+        expect(mockNextHandler).toHaveBeenCalledWith(sampleConversationStreamDoneAtIndexEvent);
+      });
+
+      test('Turn done event', async () => {
+        subs.onCreateAssistantResponseChatBot.next({
+          data: {
+            onCreateAssistantResponseChatBot: {
+              ...sampleConversationStreamTurnDoneEvent,
+            },
+          },
+        });
+
+        await pause(1);
+        expect(mockNextHandler).toHaveBeenCalledWith(sampleConversationStreamTurnDoneEvent);
+      });
+
+      test('Error event', async () => {
+        subs.onCreateAssistantResponseChatBot.next({
+          data: {
+            onCreateAssistantResponseChatBot: {
+              ...sampleConversationStreamErrorEvent,
+            },
+          },
+        });
+
+        await pause(1);
+        expect(mockErrorHandler).toHaveBeenCalledWith(sampleConversationStreamErrorEvent);
+      });
+    });
+
     test('Subscribe to messages', async () => {
       // #region mocking
       const { spy, subSpy, subs, generateClient } = mockedGenerateClient([
@@ -485,7 +685,7 @@ describe('AI Conversation Routes', () => {
           },
         },
       ]);
-      const mockHandler = jest.fn();
+      const mockNextHandler = jest.fn();
       // simulated amplifyconfiguration.json
       const config = await buildAmplifyConfig(schema);
       // #endregion mocking
@@ -499,15 +699,17 @@ describe('AI Conversation Routes', () => {
         id: sampleConversation.id,
       });
       // subscribe to messages
-      conversation?.onMessage((message) => {
-        mockHandler(message);
+      conversation?.onStreamEvent({
+        next: (streamEvent) => {
+          mockNextHandler(streamEvent);
+        },
+        error: () => { },
       });
 
       subs.onCreateAssistantResponseChatBot.next({
         data: {
           onCreateAssistantResponseChatBot: {
-            ...sampleConversationMessage1,
-            role: 'assistant',
+            ...sampleConversationStreamTextEvent,
           },
         },
       });
@@ -517,10 +719,15 @@ describe('AI Conversation Routes', () => {
       // #region assertions
       expect(optionsAndHeaders(spy)).toMatchSnapshot();
       expect(subOptionsAndHeaders(subSpy)).toMatchSnapshot();
-      expect(mockHandler).toHaveBeenCalledWith({
-        ...sampleConversationMessage1,
-        role: 'assistant',
-      });
+      const {
+        contentBlockText,
+        ...rest
+      } = sampleConversationStreamTextEvent;
+      const expectedConversationStreamEvent = {
+        text: contentBlockText,
+        ...rest,
+      };
+      expect(mockNextHandler).toHaveBeenCalledWith(expectedConversationStreamEvent);
       // #endregion assertions
     });